% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nchar.R
\name{ansi_nchar}
\alias{ansi_nchar}
\title{Number of characters or bytes in an ANSI substring}
\usage{
ansi_nchar(x, ...)
}
\arguments{
\item{x}{character vector, or a vector to be coerced to a character
vector. Giving a factor is an error.}

\item{...}{Extra arguments are passed to \code{\link[base:nchar]{base::nchar()}}.}
}
\description{
This is the ANSI-aware version of the \code{\link[base:nchar]{base::nchar()}} function.
It simply drops all ANSI styles and calls \code{\link[base:nchar]{base::nchar()}}.
}
\examples{
str <- crayon::red("I am red")

# Has ANSI markup if your terminal supports it
str

# It prints in red if your terminal supports it
cat(str)

# But base::nchar() also counts the ANSI control characters,
# which is not what we want
nchar(str)

# ansi_nchar is better:
ansi_nchar(str)

# This is the same
nchar(crayon::strip_style(str))
}
