% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strwrap.R
\name{ansi_strwrap}
\alias{ansi_strwrap}
\title{Wrap Character Strings to Format Paragraphs - ANSI Aware}
\usage{
ansi_strwrap(x, width = 0.9 * getOption("width"), indent = 0, exdent = 0,
  prefix = "", simplify = TRUE, initial = prefix)
}
\arguments{
\item{x:}{a character vector, or an object which can be converted to a
character vector by \sQuote{as.character}.}

\item{width:}{a positive integer giving the target column for wrapping
lines in the output.}

\item{indent:}{a non-negative integer giving the indentation of the first
line in a paragraph.}

\item{exdent:}{a non-negative integer specifying the indentation of
subsequent lines in paragraphs.}

\item{prefix:}{a character string to be used as prefix for each line
except the first, for which \sQuote{initial} is used.}

\item{initial:}{a character string to be used as prefix for the first line}

\item{simplify:}{a logical.  If \sQuote{TRUE}, the result is a single
character vector of line text; otherwise, it is a list of the same length
as \code{x} the elements of which are character vectors of line text
obtained from the corresponding element of \code{x}.  (Hence, the result in
the former case is obtained by unlisting that of the latter.)}
}
\description{
This is an ANSI-aware copy of \code{base::strwrap}.  The source code is a
copied and lightly modified version of that function.
}
\details{
copy stwrap docs here
}
