% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substring.R
\name{ansi_substr1}
\alias{ansi_substr1}
\title{Substring, for a single character}
\usage{
ansi_substr1(x, start, stop)
}
\arguments{
\item{x}{Length one character vector.}

\item{start}{Length one numeric or integer.}

\item{stop}{Length one numeric or integer.}
}
\value{
Length one character, the substring.
}
\description{
We use the ANSI map of the string, created by \code{\link[=make_ansi_map1]{make_ansi_map1()}}.
The result is put together from three pieces:
\itemize{
\item The ANSI start tags that are active at the beginning of the substring.
\item The substring that is selected, including all ANSI tags in it.
\item The ANSI end tags that are active at the end of the substring.
}
}
\keyword{internal}
