% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{ansi_to_html}
\alias{ansi_to_html}
\title{Convert ANSI highlighted strings to HTML}
\usage{
ansi_to_html(text, fullpage = TRUE, collapse = TRUE)
}
\arguments{
\item{text}{Character vector of ANSI highlighted strings.}

\item{fullpage}{Whether to include a minimal HTML header and footer
in the output.}

\item{collapse}{Whether to collapse the character vector to a single
string before the conversion.}
}
\value{
Character vector of HTML text.
}
\description{
Convert ANSI highlighted strings to HTML
}
