% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{make_ansi_map1}
\alias{make_ansi_map1}
\alias{make_ansi_map}
\title{Create a map of the ANSI tags of a single string}
\usage{
make_ansi_map1(str)

make_ansi_map(str)
}
\arguments{
\item{str}{A string scalar.}
}
\value{
A list with two components. \code{map} is a data frame with four
columns: \code{start}, \code{end}, \code{open}, \code{close}. It has one row for each
single ANSI markup. In other words, it has one row for each ANSI start
tag. The \code{start} column contains the start positions (in raw
coordinates) of the markup, the \code{end} column has the end positions.
The \code{open} and \code{close} columns contain the full ANSI open and close
tags.

\code{shifts} is a three column matrix that can be used to convert between
raw and ANSI coordinates. It is created by \code{\link[=make_shifts1]{make_shifts1()}}, see that
for the actual format.
}
\description{
This map can be then used in various string operations, e.g.
\code{\link[=ansi_substr]{ansi_substr()}}, etc.
}
\details{
\code{make_ansi_map1()} works for a string scalar, \code{make_ansi_map()} does
the same for all strings in a character vector.
}
\keyword{internal}
