% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{make_shifts1}
\alias{make_shifts1}
\title{Create the map for raw <-> ANSI coordinate conversion}
\usage{
make_shifts1(re)
}
\arguments{
\item{re}{The positions of the ANSI tags, as the output of the
\code{\link[rematch2:re_exec_all]{rematch2::re_exec_all()}} function.}
}
\value{
A three column matrix.
First column is the raw coordinates of the tags, second column is
ansi coordinates, third column is the offset between them.
}
\description{
The result of this function can be used with \code{\link[=map_raw_to_ansi1]{map_raw_to_ansi1()}} and
\code{\link[=map_ansi_to_raw1]{map_ansi_to_raw1()}} do the actual conversion between them.
}
\keyword{internal}
