% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_ansi_to_raw1}
\alias{map_ansi_to_raw1}
\title{Map ANSI string positions to raw positions}
\usage{
map_ansi_to_raw1(map, ansi)
}
\arguments{
\item{map}{The map matrix, created with \code{\link[=make_ansi_map1]{make_ansi_map1()}}.}

\item{ansi}{An integer or numeric scalar, the ANSI position to convert.}
}
\value{
A numeric scalar, the raw position corresponding to the
ANSI position \code{ansi}.
}
\description{
\code{\link[=map_raw_to_ansi1]{map_raw_to_ansi1()}} does the opposite convertion, from raw positions
to ANSI positions.
}
\keyword{internal}
