% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_raw_to_ansi1}
\alias{map_raw_to_ansi1}
\title{Map raw string positions to ANSI}
\usage{
map_raw_to_ansi1(map, raw)
}
\arguments{
\item{map}{The map matrix, created with \code{\link[=make_shifts1]{make_shifts1()}} or
a component of \code{\link[=make_ansi_map1]{make_ansi_map1()}} or \code{\link[=make_ansi_map]{make_ansi_map()}}.}

\item{raw}{An integer or numeric scalar, the raw position to convert.}
}
\value{
A numeric scalar, the ANSI position corresponding to the
raw position \code{raw}.
}
\description{
\code{\link[=map_ansi_to_raw1]{map_ansi_to_raw1()}} does the opposite convertion, from ANSI positions
to raw positions.
}
\keyword{internal}
