% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{myseq}
\alias{myseq}
\alias{\%:\%}
\title{Safe sequence operator}
\usage{
myseq(from, to, by = 1)

from \%:\% to
}
\arguments{
\item{from}{Start of sequence.}

\item{to}{End of sequence.}

\item{by}{Step size. Can be negative, and then \code{to} must not be more
than \code{from}.}
}
\value{
An integer vector.
}
\description{
This is a safer version of \code{\link[base:seq]{base::seq()}}. In particular, it avoids
creating backwards sequences, unless the \code{by} parameter is
explicitly negative.
}
\details{
\code{\%:\%} is the operator version of \code{myseq}.
}
\keyword{internal}
