% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ ansistrings }
\alias{ ansistrings }
\alias{ ansistrings }
\title{ Manipulation of 'ANSI' colored strings }
\description{
String manipulation functions for 'ANSI' colored strings. Creates substrings,
  split strings, wraps strings, and keeps the coloring proper.
}
