% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex.R
\name{re_ansi}
\alias{re_ansi}
\title{Create a regular expression that matches ANSI tags}
\usage{
re_ansi()
}
\value{
A character scalar, the regular expression.
}
\description{
It uses groups to distinguish between start and end tags, and
it also keeps the actual command string within the ANSI tag. E.g.
\code{"1"} means a \emph{bold} start tag.
}
\keyword{internal}
