% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex.R
\name{re_closes}
\alias{re_closes}
\title{Which ANSI end tag closes which ANSI start tag?}
\usage{
re_closes()
}
\value{
A named list, the names are the ANSI end tag command strings.
The list contains character vectors, specifying the start tags that
are closed by each end tag. Most end tags close a single start tag,
but some close multiple ones, and the \emph{reset} end tag closes all
start tags.
}
\description{
Create a map that answers this question.
}
\keyword{internal}
